Introduction to closures
This commit introduces closures which are executable blocks of code. It shows how we can
parameters to closures, how the last statement of a closure is it's return value and how
"it" is a default parameter when no other parameter is defined. It also shows how
closures can bind variables in the scope in which they are defined, which a very
important characteristic of closures and how we can send closures as parameters to
other closures.